Move CandidateMerge into it's own files.
